+++ internal/docs_test.go
@@ -259,6 +259,11 @@ func Test_GenerateAPIDocs(t *testing.T) {
 				},
 				Shared: []string{
 					"shared/v1",
+					"scheduler/v1beta1",
+					"scheduler/v1beta1/container",
+					"scheduler/v1beta1/container/resources",
+					"scheduler/v1beta1/pod",
+					"scheduler/v1beta1/pod/resources",
 				},
 			},
 		},