+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - (Maintenance) Update Go to 1.22.3
 - (Feature) (Analytics) GAE Integration
 - (Feature) (Analytics) Enable TLS and Service
+- (Feature) (ML) Unify Integration Sidecar
 
 ## [1.2.40](https://github.com/arangodb/kube-arangodb/tree/1.2.40) (2024-04-10)
 - (Feature) Add Core fields to the Scheduler Container Spec